refactor: created internal dir for private libraries and rename main dir from src to cmd
