+++ Dockerfile
@@ -1,5 +1,5 @@
 # Builder stage
-FROM golang:1.19.5-alpine3.17 as BUILDER
+FROM golang:1.21.0-alpine3.17 as BUILDER
 # install gcc and musl-dev for compiling go code
 RUN apk add gcc musl-dev
 
@@ -10,9 +10,10 @@ COPY go.sum .
 
 RUN go mod download
 
-COPY src ./src
+COPY cmd ./cmd
+COPY internal ./internal
 
-RUN go build ./src/main.go
+RUN go build ./cmd/main.go
 
 # Final Stage
 FROM alpine:3.17
@@ -20,10 +21,10 @@ FROM alpine:3.17
 WORKDIR /tiddi
 
 # Copy sample frontend
-COPY --from=BUILDER /tiddi/src/frontend ./src/frontend
+COPY --from=BUILDER /tiddi/cmd/frontend ./cmd/frontend
 
 COPY --from=BUILDER /tiddi/main .
 
 EXPOSE 5656
 
-CMD ["./main"]
+ENTRYPOINT ["./main"]
