+++ cmd/api/cors.go
